EES-2348 Move ContentDbUtils and StatisticsDbUtils into Content.Model.Tests and Data.Model.Tests
