+++ src/GovUk.Education.ExploreEducationStatistics.Admin.Tests/Services/DbUtils.cs
@@ -4,7 +4,7 @@ using GovUk.Education.ExploreEducationStatistics.Content.Model.Database;
 using IdentityServer4.EntityFramework.Options;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Options;
-using static GovUk.Education.ExploreEducationStatistics.Common.Tests.Utils.ContentDbUtils;
+using static GovUk.Education.ExploreEducationStatistics.Content.Model.Tests.Utils.ContentDbUtils;
 
 namespace GovUk.Education.ExploreEducationStatistics.Admin.Tests.Services
 {
