+++ src/GovUk.Education.ExploreEducationStatistics.Admin.Tests/Services/ReleaseSubjectServiceTests.cs
@@ -6,6 +6,7 @@ using GovUk.Education.ExploreEducationStatistics.Data.Model;
 using GovUk.Education.ExploreEducationStatistics.Data.Model.Database;
 using GovUk.Education.ExploreEducationStatistics.Data.Model.Repository;
 using GovUk.Education.ExploreEducationStatistics.Data.Model.Repository.Interfaces;
+using GovUk.Education.ExploreEducationStatistics.Data.Model.Tests.Utils;
 using Microsoft.EntityFrameworkCore;
 using Moq;
 using Xunit;
